requirements.txt
# Core Dependencies
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Free LLM Options
google-generativeai>=0.3.0
langchain-google-genai>=0.0.5
ollama>=0.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
PyPDF2>=3.0.0
python-docx>=1.1.0

# API and Web
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Utilities
rich>=13.7.0
tiktoken>=0.5.2
